Split inbound::server stack into distinct modules
