+++ linkerd/app/outbound/src/server.rs
@@ -1,179 +1,11 @@
-#![allow(clippy::too_many_arguments)]
-
-use crate::{http, stack_labels, target::ShouldResolve, tcp, trace_labels, Config};
+use crate::{http, stack_labels, target::ShouldResolve, tcp};
 use linkerd_app_core::{
     config::{ProxyConfig, ServerConfig},
-    detect, discovery_rejected, drain, errors, io, metrics,
-    opencensus::proto::trace::v1 as oc,
-    profiles,
-    proxy::{api_resolve::Metadata, core::resolve::Resolve},
-    spans::SpanConverter,
-    svc,
-    transport::{listen, metrics::SensorIo},
-    Addr, Error, IpMatch, TraceContext,
+    detect, discovery_rejected, drain, io, metrics, profiles, svc, Addr, Error, IpMatch,
 };
-use std::net::SocketAddr;
-use tokio::sync::mpsc;
 use tracing::debug_span;
 
-pub fn stack<R, P, C, H, HSvc, I>(
-    config: &Config,
-    profiles: P,
-    resolve: R,
-    tcp_connect: C,
-    http_router: H,
-    metrics: metrics::Proxy,
-    span_sink: Option<mpsc::Sender<oc::Span>>,
-    drain: drain::Watch,
-) -> impl svc::NewService<
-    listen::Addrs,
-    Service = impl svc::Service<I, Response = (), Error = Error, Future = impl Send>,
->
-where
-    I: io::AsyncRead + io::AsyncWrite + io::PeerAddr + std::fmt::Debug + Send + Unpin + 'static,
-    R: Resolve<Addr, Endpoint = Metadata, Error = Error> + Clone + Send + 'static,
-    R::Resolution: Send,
-    R::Future: Send + Unpin,
-    C: svc::Service<tcp::Endpoint> + Clone + Send + 'static,
-    C::Response: tokio::io::AsyncRead + tokio::io::AsyncWrite + Send + Unpin,
-    C::Error: Into<Error>,
-    C::Future: Send,
-    H: svc::NewService<http::Logical, Service = HSvc> + Clone + Send + 'static,
-    HSvc: svc::Service<http::Request<http::BoxBody>, Response = http::Response<http::BoxBody>>
-        + Send
-        + 'static,
-    HSvc::Error: Into<Error>,
-    HSvc::Future: Send,
-    P: profiles::GetProfile<SocketAddr> + Clone + Send + 'static,
-    P::Future: Send,
-    P::Error: Send,
-{
-    let tcp = tcp::balance::stack(&config.proxy, tcp_connect, resolve, &metrics, drain.clone());
-    let http = http(&config.proxy, &metrics, span_sink, http_router);
-    let accept = dispatch(&config.proxy, &metrics, drain, tcp, http);
-    discover(config, &metrics, profiles, accept)
-}
-
-/// Wraps an `N`-typed TCP stack with caching.
-///
-/// Services are cached as long as they are retained by the caller (usually
-/// core::serve) and are evicted from the cache when they have been dropped for
-/// `config.cache_max_idle_age` with no new acquisitions.
-pub fn discover<I, N, NSvc, P>(
-    config: &Config,
-    metrics: &metrics::Proxy,
-    profiles: P,
-    stack: N,
-) -> impl svc::NewService<
-    listen::Addrs,
-    Service = impl svc::Service<I, Response = (), Error = Error, Future = impl Send>,
->
-where
-    I: io::AsyncRead + io::AsyncWrite + io::PeerAddr + std::fmt::Debug + Send + Unpin + 'static,
-    N: svc::NewService<tcp::Logical, Service = NSvc> + Clone + Send + 'static,
-    NSvc: svc::Service<SensorIo<I>, Response = ()> + Send + 'static,
-    NSvc::Error: Into<Error>,
-    NSvc::Future: Send,
-    P: profiles::GetProfile<SocketAddr> + Clone + Send + 'static,
-    P::Future: Send,
-    P::Error: Send,
-{
-    svc::stack(stack)
-        .push_on_response(svc::MapErrLayer::new(Into::into))
-        .check_new::<tcp::Logical>()
-        .check_new_service::<tcp::Logical, SensorIo<I>>()
-        .push_map_target(tcp::Logical::from)
-        .push(profiles::discover::layer(
-            profiles,
-            AllowProfile(config.allow_discovery.clone().into()),
-        ))
-        .check_new_service::<tcp::Accept, SensorIo<I>>()
-        .push_on_response(
-            svc::layers()
-                // If the traffic split is empty/unavailable, eagerly fail
-                // requests requests. When the split is in failfast, spawn
-                // the service in a background task so it becomes ready without
-                // new requests.
-                .push(svc::layer::mk(svc::SpawnReady::new))
-                .push(metrics.stack.layer(crate::stack_labels("tcp", "server")))
-                .push(svc::FailFast::layer(
-                    "TCP Server",
-                    config.proxy.dispatch_timeout,
-                ))
-                .push_spawn_buffer(config.proxy.buffer_capacity),
-        )
-        .check_new_service::<tcp::Accept, SensorIo<I>>()
-        .push(metrics.transport.layer_accept())
-        .push_cache(config.proxy.cache_max_idle_age)
-        .check_new_service::<tcp::Accept, I>()
-        .push_map_target(tcp::Accept::from)
-        .check_new_service::<listen::Addrs, I>()
-        .into_inner()
-}
-
-pub fn http<H, HSvc>(
-    config: &ProxyConfig,
-    metrics: &metrics::Proxy,
-    span_sink: Option<mpsc::Sender<oc::Span>>,
-    http: H,
-) -> impl svc::NewService<
-    http::Logical,
-    Service = impl svc::Service<
-        http::Request<http::BoxBody>,
-        Response = http::Response<http::BoxBody>,
-        Error = Error,
-        Future = impl Send,
-    > + Clone,
-> + Clone
-where
-    H: svc::NewService<http::Logical, Service = HSvc> + Clone + Send + 'static,
-    HSvc: svc::Service<http::Request<http::BoxBody>, Response = http::Response<http::BoxBody>>
-        + Send
-        + 'static,
-    HSvc::Error: Into<Error>,
-    HSvc::Future: Send,
-{
-    let ProxyConfig {
-        dispatch_timeout,
-        max_in_flight_requests,
-        buffer_capacity,
-        ..
-    } = config.clone();
-
-    svc::stack(http)
-        .check_new_service::<http::Logical, _>()
-        .push_on_response(
-            svc::layers()
-                .push(http::BoxRequest::layer())
-                // Limit the number of in-flight requests. When the proxy is
-                // at capacity, go into failfast after a dispatch timeout. If
-                // the router is unavailable, then spawn the service on a
-                // background task to ensure it becomes ready without new
-                // requests being processed.
-                .push(svc::layer::mk(svc::SpawnReady::new))
-                .push(svc::ConcurrencyLimit::layer(max_in_flight_requests))
-                .push(svc::FailFast::layer("HTTP Server", dispatch_timeout))
-                .push_spawn_buffer(buffer_capacity)
-                .push(metrics.http_errors.clone())
-                // Synthesizes responses for proxy errors.
-                .push(errors::layer())
-                // Initiates OpenCensus tracing.
-                .push(TraceContext::layer(span_sink.map(|span_sink| {
-                    SpanConverter::server(span_sink, trace_labels())
-                })))
-                .push(http::BoxResponse::layer()),
-        )
-        // Convert origin form HTTP/1 URIs to absolute form for Hyper's
-        // `Client`.
-        .push(http::NewNormalizeUri::layer())
-        // Record when a HTTP/1 URI originated in absolute form
-        .push_on_response(http::normalize_uri::MarkAbsoluteForm::layer())
-        .instrument(|l: &http::Logical| debug_span!("http", v = %l.protocol))
-        .check_new_service::<http::Logical, http::Request<http::BoxBody>>()
-        .into_inner()
-}
-
-pub fn dispatch<T, TSvc, H, HSvc, I>(
+pub fn stack<T, TSvc, H, HSvc, I>(
     config: &ProxyConfig,
     metrics: &metrics::Proxy,
     drain: drain::Watch,
